autopilot: add errors.go
